api testing suite part 2 #1102 (#1135)
* Added unstaking and reosurce stake endowment #1102

* Suite structure refactoring part 1 #1102

Split verifier on two files: GetAccount verifier and
TransactionExecutionVerifier. Verification functions made as stand
alone, not as class methods

* Suite structure refactoring part 2 #1102

The suite files splitted on tools and framwork subdirs. Test cases files
splitted on verifiation and case subdirs

* Renamed verificators #1102

The suite files splitted on tools and framwork subdirs. Test cases files
splitted on verifiation and case subdirs

* Added transfer stake tests#1102

staking for another user
delegatebw with transfer

* BootSequenceTest extracted from CleosWalletTest #1102

Added support to run several test files on a one Nodeos instance
Splitted the CleosWalletTest on  the GetAccountTest and BootSequence
tests.
Minor: Added a eprint function to print output to stderr, as the
unittest suite prints it output to the stderr